requirements.txt
numpy
scipy
scikit-learn
pandas
matplotlib
seaborn
pyclustering
pytorch